Clean up download tests and require absolute paths for downloads
