+++ jasper_test.go
@@ -32,7 +32,10 @@ func getPortNumber() int {
 
 }
 
-const taskTimeout = 5 * time.Second
+const (
+	taskTimeout     = 5 * time.Second
+	longTaskTimeout = 100 * time.Second
+)
 
 // this file contains tools and constants used throughout the test
 // suite.
