+++ jrpc/client_test.go
@@ -214,7 +214,7 @@ func TestJRPCManager(t *testing.T) {
 				// "": func(ctx context.Context, t *testing.T, manager jasper.Manager) {},
 			} {
 				t.Run(name, func(t *testing.T) {
-					tctx, cancel := context.WithTimeout(ctx, longTaskTimeout)
+					tctx, cancel := context.WithTimeout(ctx, taskTimeout)
 					defer cancel()
 					test(tctx, t, factory(tctx, t))
 				})
